+++ api/src/mutations/signUp.ts
@@ -2,7 +2,6 @@ import HttpStatus from 'http-status-codes';
 
 import { buildError } from '../utils/errors';
 import Context from '../models/Context';
-
 import * as userService from '../service/user';
 
 export const signUp = async (
